Factor out administration notes component
