+++ frontend/src/views/InstructionPage/index.vue
@@ -95,26 +95,7 @@
               <v-icon name="ri-pencil-fill"></v-icon>
               Notes à destination de l'administration
             </h6>
-            <div class="text-left mb-4 sm:mb-0 sm:flex sm:gap-8">
-              <DsfrInputGroup>
-                <DsfrInput
-                  v-model="privateNotesInstruction"
-                  is-textarea
-                  label-visible
-                  label="Notes de l'instruction"
-                  @update:modelValue="saveComment"
-                />
-              </DsfrInputGroup>
-              <DsfrInputGroup>
-                <DsfrInput
-                  :disabled="true"
-                  v-model="privateNotesVisa"
-                  is-textarea
-                  label-visible
-                  label="Notes du visa"
-                />
-              </DsfrInputGroup>
-            </div>
+            <AdministrationNotes class="mb-4" v-model="declaration" :disableVisaNotes="true" />
           </template>
         </TabStepper>
       </div>
@@ -127,13 +108,14 @@ import TabStepper from "@/components/TabStepper"
 import { useRootStore } from "@/stores/root"
 import { storeToRefs } from "pinia"
 import { onMounted, computed, ref } from "vue"
-import { useFetch, useDebounceFn } from "@vueuse/core"
+import { useFetch } from "@vueuse/core"
 import { handleError } from "@/utils/error-handling"
 import ProgressSpinner from "@/components/ProgressSpinner"
 import DeclarationSummary from "@/components/DeclarationSummary"
 import IdentityTab from "@/components/IdentityTab"
 import HistoryTab from "@/components/HistoryTab"
 import DecisionTab from "./DecisionTab"
+import AdministrationNotes from "@/components/AdministrationNotes"
 import { headers } from "@/utils/data-fetching"
 import DeclarationAlert from "@/components/DeclarationAlert"
 import { tabTitles } from "@/utils/mappings"
@@ -166,8 +148,7 @@ const {
   execute: executeDeclarationFetch,
   isFetching: isFetchingDeclaration,
 } = useFetch(`/api/v1/declarations/${props.declarationId}`, { immediate: false }).get().json()
-const privateNotesInstruction = ref(declaration.value?.privateNotesInstruction || "")
-const privateNotesVisa = ref(declaration.value?.privateNotesVisa || "")
+
 const {
   response: declarantResponse,
   data: declarant,
@@ -192,22 +173,9 @@ const {
   .get()
   .json()
 
-// Sauvegarde du commentaire privé
-const saveComment = useDebounceFn(async () => {
-  const { response } = await useFetch(() => `/api/v1/declarations/${declaration.value?.id}`, {
-    headers: headers(),
-  })
-    .patch({ privateNotesInstruction: privateNotesInstruction.value })
-    .json()
-  handleError(response)
-}, 600)
-
 onMounted(async () => {
   await executeDeclarationFetch()
   handleError(declarationResponse)
-
-  privateNotesInstruction.value = declaration.value?.privateNotesInstruction || ""
-  privateNotesVisa.value = declaration.value?.privateNotesVisa || ""
 
   // Si on arrive à cette page avec une déclaration déjà assignée à quelqun.e mais en état
   // AWAITING_INSTRUCTION, on la passe directement à ONGOING_INSTRUCTION.
